Split blitz-shell out of dioxus-native crate
