+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
   "packages/blitz-net",
   "packages/blitz-renderer-vello",
   "packages/blitz-shell",
+  "packages/dioxus-native",
   "packages/stylo_taffy",
   "apps/wpt",
   "apps/readme",
@@ -112,6 +113,7 @@ blitz-net = { path = "./packages/blitz-net" }
 blitz-traits = { path = "./packages/blitz-traits" }
 blitz-renderer-vello = { path = "./packages/blitz-renderer-vello" }
 blitz-shell = { path = "./packages/blitz-shell", features = ["tracing"] }
+dioxus-native = { path = "./packages/dioxus-native", features = ["tracing"] }
 png = { version = "0.17" }
 dioxus = { workspace = true }
 euclid = { version = "0.22", features = ["serde"] }
