+++ apps/readme/src/main.rs
@@ -6,7 +6,9 @@ use blitz_traits::net::SharedCallback;
 use markdown::{markdown_to_html, BLITZ_MD_STYLES, GITHUB_MD_STYLES};
 use reqwest::header::HeaderName;
 
-use blitz_shell::{create_default_event_loop, BlitzApplication, WindowConfig, WinitNetCallback};
+use blitz_shell::{
+    create_default_event_loop, BlitzApplication, BlitzShellNetCallback, WindowConfig,
+};
 use std::env::current_dir;
 use std::fs;
 use std::path::{Path, PathBuf};
@@ -42,7 +44,7 @@ fn main() {
 
     // println!("{html}");
 
-    let net_callback = Arc::new(WinitNetCallback::new(proxy));
+    let net_callback = Arc::new(BlitzShellNetCallback::new(proxy));
     let net_provider = Arc::new(Provider::new(
         Handle::current(),
         Arc::clone(&net_callback) as SharedCallback<Resource>,
