+++ examples/bare_style.rs
@@ -33,7 +33,7 @@ fn root() -> Element {
 }
 
 fn main() {
-    blitz_shell::launch(root);
+    dioxus_native::launch(root);
 
     // let document = blitz_dom::Document::new();
