+++ examples/flex_order.rs
@@ -7,7 +7,7 @@ Servo doesn't have:
 use dioxus::prelude::*;
 
 fn main() {
-    blitz_shell::launch(app);
+    dioxus_native::launch(app);
 }
 
 fn app() -> Element {
