+++ packages/dioxus-native/Cargo.toml
@@ -1,14 +1,15 @@
 [package]
-name = "blitz-shell"
+name = "dioxus-native"
 version = "0.0.0"
 license.workspace = true
 edition = "2021"
 
 [features]
-accessibility = ["dep:accesskit", "dep:accesskit_winit"]
-menu = ["dep:muda"]
-tracing = ["dep:tracing"]
-default = ["accessibility", "menu", "tracing"]
+default = ["accessibility", "hot-reload", "menu", "tracing"]
+accessibility = ["blitz-shell/accessibility"]
+menu = ["blitz-shell/menu"]
+tracing = ["dep:tracing", "blitz-shell/tracing"]
+hot-reload = ["dep:dioxus-cli-config", "dep:dioxus-devtools"]
 
 [dependencies]
 # Blitz dependencies
@@ -16,27 +17,27 @@ blitz-renderer-vello = { path = "../blitz-renderer-vello" }
 blitz-dom = { path = "../blitz-dom" }
 blitz-net = { path = "../blitz-net" }
 blitz-traits = { path = "../blitz-traits" }
+blitz-shell = { path = "../blitz-shell" }
 
 # Servo dependencies
 style = { workspace = true }
 
+# DioxusLabs dependencies
+dioxus = { workspace = true }
+dioxus-cli-config = { workspace = true, optional = true }
+dioxus-devtools = { workspace = true, optional = true }
+
 # Windowing & Input
 winit = { workspace = true }
-muda = { workspace = true, features = ["serde"], optional = true }
-accesskit = { workspace = true, optional = true }
-accesskit_winit = {workspace = true, optional = true }
+keyboard-types = { workspace = true }
 
 # IO & Networking
-url = { workspace = true, features = ["serde"] }
 tokio = { workspace = true, features = ["full"] }
-ureq = { workspace = true }
 
 # Other dependencies
 tracing = { workspace = true, optional = true }
+rustc-hash = { workspace = true }
 futures-util = { workspace = true }
-
-[target.'cfg(target_os = "android")'.dependencies]
-android-activity = { version = "0.6.0", features = ["native-activity"] }
 
 [package.metadata.docs.rs]
 all-features = true
